Add files via upload
AI Fairness 360 is an open source toolkit that will help us to examine, report, and mitigate discrimination and bias in machine learning models throughout the AI application life-cycle. Containing over 70 fairness metrics and 10 state-of-the-art bias mitigation algorithms developed by the research community, it is designed to translate algorithmic research from the lab into the actual practice of domains as wide-ranging as finance, human capital management, healthcare, and education.